[Refactor] Post와 Notice 분리
